requirements.txt
aiogram>=3.0
sqlalchemy[asyncio]>=2.0
aiosqlite
pydantic-settings
python-dotenv
watchfiles
openai
